% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{addMinMaxCols}
\alias{addMinMaxCols}
\title{FUNCTION_TITLE}
\usage{
addMinMaxCols(dt, num_cols)
}
\arguments{
\item{dt}{PARAM_DESCRIPTION}

\item{num_cols}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
