% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{cleanPlayerData}
\alias{cleanPlayerData}
\title{FUNCTION_TITLE}
\usage{
cleanPlayerData(dt, dummies)
}
\arguments{
\item{dt}{PARAM_DESCRIPTION}

\item{dummies}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
