% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{extractBpsData}
\alias{extractBpsData}
\title{FUNCTION_TITLE}
\usage{
extractBpsData(fpl_player_hist, fpl_season_json)
}
\arguments{
\item{fpl_player_hist}{PARAM_DESCRIPTION}

\item{fpl_season_json}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
