% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_2017_18_player_history.R
\docType{data}
\name{fpl_2017_18_player_history}
\alias{fpl_2017_18_player_history}
\title{2017/18 player and fixture data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10448 rows and 50 columns.}
\source{
\url{https://github.com/vaastav/Fantasy-Premier-League}
}
\usage{
fpl_2017_18_player_history
}
\description{
2017/18 player and fixture data
}
\examples{
# see this function for details on how it was packaged
storePackageData()
}
\keyword{datasets}
