% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_player_hist.R
\docType{data}
\name{fpl_player_hist}
\alias{fpl_player_hist}
\title{Player historical FTP data}
\format{An object of class \code{character} of length 1.}
\source{
\url{https://fantasy.premierleague.com/drf/element-summary}
}
\usage{
fpl_player_hist
}
\description{
input from a \code{\link{fpl_season_json}} list object.
Loops through the ids, and collects all the historical season data for 
these players
}
\examples{
fpl_season_json <- pullFplSeasonJson() 
}
\keyword{datasets}
