% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_season_json.R
\docType{data}
\name{fpl_season_json}
\alias{fpl_season_json}
\title{JSON extract of latest seasons fpl data}
\format{An object of class \code{list} of length 13.}
\source{
\url{https://fantasy.premierleague.com/drf/bootstrap-static}
}
\usage{
fpl_season_json
}
\description{
data extract in list output from a static url
}
\examples{
fpl_season_json <- pullFplSeasonJson() 
fpl_player_hist <- pullFplHistoryDt(fpl_season_json)
}
\keyword{datasets}
