% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{playerLookup}
\alias{playerLookup}
\title{FUNCTION_TITLE}
\usage{
playerLookup(fpl_season_json)
}
\arguments{
\item{fpl_season_json}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringi]{stri_trans_general}}
}
